Built in quit

+++ signals.cpp
@@ -11,11 +11,19 @@ void ctrlZHandler(int sig_num) {
 }
 
 void ctrlCHandler(int sig_num) {
-  // TODO: Add your implementation
+  std::cout << "smash: got ctrl-C\n";
+  // Get current foreground command (if none, then fgCommand will be smash pid)
   SmallShell& smash = SmallShell::getInstance(); // Same smash instance as in main
-  std::cout << "Got C signal\n"; // TODO: delete
   pid_t fgCommand = smash.returnFgPid();
-  kill(fgCommand, SIGKILL);
+
+  if (fgCommand != getpid()) {
+    // IF there is a child (in foreground) then kill fgCommand (child)
+    kill(fgCommand, SIGKILL);
+    std::cout << "smash: process " << std::to_string(fgCommand) <<" was killed\n";
+  } else {
+    // TODO: Check how the "smash>" should be printed, maybe I should throw; and catch; in main instead of continuing on std::getline()
+  }
+  
 }
 
 void alarmHandler(int sig_num) {
